feat: Specify null encoding for JSON rows
When a field is annotated with the 'json' attribute, require an
additional attribute, either 'null_is_json' or 'null_is_sql', to be
supplied to indicate how null values should be stored in the database.

The behaviour of lsor prior to this commit has been as follows:
- If the 'json' attribute is specified, store null values as JSON null
- If the 'json' attribute is *not* specified, store null values as SQL
  null

The behaviour now is:
- If the 'json' attribute is specified, use the specified approach
- If the 'json' attribute is *not* specified, store null values as SQL
  null (as before)

For all fields marked with the 'json' attribute, we now accept either
kind of null, instead of returning an error if a SQL null was found
where a JSON null was expected.